heroku run:detached --tail
Tails the logs of the run processes.

Restructured the log display logic into a helper class.

Inspired by a convo with @csquared and (maybe) others.
