+++ lib/heroku/command/logs.rb
@@ -1,4 +1,5 @@
 require "heroku/command/base"
+require "heroku/helpers/log_displayer"
 
 # display logs for an app
 #
@@ -28,25 +29,8 @@ class Heroku::Command::Logs < Heroku::Command::Base
     opts << "ps=#{URI.encode(options[:ps])}"         if options[:ps]
     opts << "source=#{URI.encode(options[:source])}" if options[:source]
 
-    @assigned_colors = {}
-    @line_start = true
-    @token = nil
-
-    heroku.read_logs(app, opts) do |chunk|
-      unless chunk.empty?
-        if STDOUT.isatty && ENV.has_key?("TERM")
-          display(colorize(chunk))
-        else
-          display(chunk)
-        end
-      end
-    end
-  rescue Errno::EPIPE
-  rescue Interrupt => interrupt
-    if STDOUT.isatty && ENV.has_key?("TERM")
-      display("\e[0m")
-    end
-    raise(interrupt)
+    log_displayer = ::Heroku::Helpers::LogDisplayer.new(heroku, app, opts)
+    log_displayer.display_logs
   end
 
   # logs:drains
@@ -58,40 +42,4 @@ class Heroku::Command::Logs < Heroku::Command::Base
     display("~ `heroku logs:drains` has been deprecated and replaced with `heroku drains`")
     Heroku::Command::Drains.new.index
   end
-
-  protected
-
-  COLORS = %w( cyan yellow green magenta red )
-  COLOR_CODES = {
-    "red"     => 31,
-    "green"   => 32,
-    "yellow"  => 33,
-    "magenta" => 35,
-    "cyan"    => 36,
-  }
-
-  def colorize(chunk)
-    lines = []
-    chunk.split("\n").map do |line|
-      if parsed_line = parse_log(line)
-        header, identifier, body = parsed_line
-        @assigned_colors[identifier] ||= COLORS[@assigned_colors.size % COLORS.size]
-        lines << [
-          "\e[#{COLOR_CODES[@assigned_colors[identifier]]}m",
-          header,
-          "\e[0m",
-          body,
-        ].join("")
-      elsif not line.empty?
-        lines << line
-      end
-    end
-    lines.join("\n")
-  end
-
-  def parse_log(log)
-    return unless parsed = log.match(/^(.*?\[(\w+)([\d\.]+)?\]:)(.*)?$/)
-    [1, 2, 4].map { |i| parsed[i] }
-  end
-
 end
